app/requirements.txt
wheel
cassandra-driver
venv-pack
nltk
pathvalidate
cqlsh
pyspark
regex
six
tqdm